Move OpenMP UniqueToken to its own file
